object.cc : Fix testcase CreateCopyDestroyObject
Testcase CreateCopyDestroyObject tries to create a copy of
object of type CKO_DATA and tries to modify it's value. This
is not allowed according to PKCS#11 specification.

As per specification only attributes marked with "8, 10, 11 and 12"
are modifiable after an object has been creates/key generated.
The only exception to these are any attributes if mentioned in
footnote of the other tables. (eg table 16 which in its footnote
specifies explicitly about the modifiable attributes). For Data Object,
attributes (Table 17) neither are the attributes marked with 8,10,11 or 12
as per Table 10, nor is it specified that these are modifiable later.
So, this suggests that attribute CKA_VALUE can't be modified.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>